fix totalCP CP/t LONG

+++ src/main/java/com/yukke9265/cobblestone_xx_compressed/blockentity/CobblestoneChemicalReactorBlockEntity.java
@@ -496,7 +496,7 @@ public class CobblestoneChemicalReactorBlockEntity extends BaseBlockEntity imple
             if (this.canProcess(recipe)) {
                 int progressStep = this.getProgressStep(recipe.getCobblestonePowerPerTick());
                 this.progress += progressStep;
-                this.storedCobblestonePower -= (long) recipe.getCobblestonePowerPerTick() * progressStep;
+                this.storedCobblestonePower -= recipe.getCobblestonePowerPerTick() * progressStep;
                 shouldTurnOn = true;
                 this.setChanged();
 
@@ -877,7 +877,7 @@ public class CobblestoneChemicalReactorBlockEntity extends BaseBlockEntity imple
         outputStack.grow(resultStack.getCount());
     }
 
-    private int getProgressStep(int cobblestonePowerPerTick) {
+    private int getProgressStep(long cobblestonePowerPerTick) {
         if (cobblestonePowerPerTick <= 0) {
             return 0;
         }
