+++ src/main/java/com/yukke9265/cobblestone_xx_compressed/blockentity/CobblestoneLaserDrillBlockEntity.java
@@ -302,7 +302,7 @@ public class CobblestoneLaserDrillBlockEntity extends BaseBlockEntity implements
             if (this.canProcess(recipe)) {
                 int progressStep = this.getProgressStep(recipe.getCobblestonePowerPerTick());
                 this.progress += progressStep;
-                this.storedCobblestonePower -= (long) recipe.getCobblestonePowerPerTick() * progressStep;
+                this.storedCobblestonePower -= recipe.getCobblestonePowerPerTick() * progressStep;
                 shouldTurnOn = true;
                 this.setChanged();
 
@@ -393,7 +393,7 @@ public class CobblestoneLaserDrillBlockEntity extends BaseBlockEntity implements
         return this.canAcceptResult(OUTPUT_SLOT_2_INDEX, recipe.getSecondResult());
     }
 
-    private int getProgressStep(int cobblestonePowerPerTick) {
+    private int getProgressStep(long cobblestonePowerPerTick) {
         if (cobblestonePowerPerTick <= 0) {
             return 0;
         }
